+++ exts/ndd.f90
@@ -22,13 +22,13 @@ module dirichlet_mod
 
 contains
 
-  subroutine initialize_dirichlet(counts, nc_)
+  subroutine initialize_dirichlet(counts, nc)
     ! set alphabet_size, n_data
     ! set n_multi, multi_z, multi
     integer(int32), intent(in) :: counts(:)
-    integer(int32), intent(in) :: nc_
+    integer(int32), intent(in) :: nc
     
-    alphabet_size = nc_
+    alphabet_size = nc
     n_data = sum(counts)
     
     call compute_multiplicities(counts)
@@ -80,43 +80,91 @@ contains
     
   end subroutine dirichlet_finalize
 
-  pure real(real64) function log_fpxa(alpha) 
+  pure real(real64) function log_pna(alpha) 
     ! log(p(n|a)) (log of) marginal probability of data given alpha
     ! computed from histogram multiplicities. Dirichlet-multinomial.
     use constants
     
     real(real64), intent(in) :: alpha
     integer(int32) :: i_
-    real(real64)   :: a(n_multi)
+    real(real64)   :: wrk(n_multi)
 
-    log_fpxa = log_gamma(n_data + one) + log_gamma(alpha * alphabet_size) & 
+    log_pna = log_gamma(n_data + one) + log_gamma(alpha * alphabet_size) & 
          - alphabet_size * log_gamma(alpha) - log_gamma(n_data + alpha * alphabet_size)
     do i_ = 1, n_multi 
-       a(i_) = multi(i_) * (log_gamma(multi_z(i_) + alpha) - log_gamma(multi_z(i_) + one))
+       wrk(i_) = multi(i_) * (log_gamma(multi_z(i_) + alpha) - log_gamma(multi_z(i_) + one))
     end do
-    log_fpxa = log_fpxa + sum(a)
+    log_pna = log_pna + sum(wrk)
     
-  end function log_fpxa
+  end function log_pna
 
-  real(real64) function hdir(alpha)
+  real(real64) function h_bayes(alpha)
     ! posterior average of the entropy given the data and alpha
     ! computed from histogram multiplicities
     use gamma_funcs, only: digamma
     use constants
     
     real(real64), intent(in) :: alpha
-    integer(int32) :: i
-    real(real64)   :: a(n_multi)
+    integer(int32) :: i_
+    real(real64)   :: wrk(n_multi)
 
-    hdir = 0.0_real64
-    do i = 1,n_multi 
-       a(i) = multi(i) * (multi_z(i) + alpha) * digamma(multi_z(i) + alpha + one) 
+    h_bayes = 0.0_real64
+    do i_ = 1,n_multi 
+       wrk(i_) = multi(i_) * (multi_z(i_) + alpha) * digamma(multi_z(i_) + alpha + one) 
     end do
-    hdir = - sum(a)
-    hdir = hdir / (n_data + alpha * alphabet_size)
-    hdir = hdir + digamma(n_data + alpha * alphabet_size + one)
+    h_bayes = - sum(wrk)
+    h_bayes = h_bayes / (n_data + alpha * alphabet_size)
+    h_bayes = h_bayes + digamma(n_data + alpha * alphabet_size + one)
 
-  end function hdir
+  end function h_bayes
+
+  elemental real(real64) function integrand(alpha, lw_max, order)
+    ! posterior average of the entropy given the data and alpha
+    ! computed from histogram multiplicities
+    use gamma_funcs, only: digamma
+    use constants
+    
+    real(real64), intent(in) :: alpha, lw_max
+    integer(int32), intent(in) :: order
+    real(real64) :: hb, lw
+    real(real64) :: lpna
+    integer(int32) :: mi, mzi
+    integer(int32) :: i_
+    real(real64)   :: wrka(n_multi),wrkb(n_multi)
+
+    lpna = log_gamma(n_data + one) &
+         + log_gamma(alpha * alphabet_size) & 
+         - alphabet_size * log_gamma(alpha) &
+         - log_gamma(n_data + alpha * alphabet_size)
+
+    do i_ = 1,n_multi
+       mzi = multi_z(i_)
+       mi = multi(i_)
+       wrka(i_) = mi * (mzi + alpha) * digamma(mzi + alpha + one)
+       wrkb(i_) = mi * (log_gamma(mzi + alpha) - log_gamma(mzi + one))
+    end do
+
+    lpna = lpna + sum(wrkb)
+    lw = log_fpa(alpha) + lpna - lw_max
+    
+    hb = - sum(wrka)
+    hb = hb / (n_data + alpha * alphabet_size)
+    hb = hb + digamma(n_data + alpha * alphabet_size + one)
+    
+    integrand = exp(lw) * hb**order
+
+  end function integrand
+
+  elemental real(real64) function log_fpa(alpha) 
+    ! prop. to p(alpha) - the prior for alpha in NSB estimator
+    use constants
+    use gamma_funcs, only: trigamma
+    
+    real(real64), intent(in) :: alpha
+    
+    log_fpa = log(alphabet_size * trigamma(alphabet_size * alpha + one) - trigamma(alpha + one))
+    
+  end function log_fpa
 
 end module dirichlet_mod
 
@@ -127,114 +175,115 @@ module nsb_mod
   real(real64) :: log_alpha1
   real(real64) :: log_alpha2
   real(real64) :: amax
-  real(real64) :: log_fwa_amax
+  real(real64) :: lw_max
 
 contains
 
-  elemental real(real64) function log_fpa(a) 
-    ! prop. to p(alpha) - the prior for alpha in NSB estimator
-    use constants
-    use gamma_funcs, only: trigamma
-    use dirichlet_mod, only: alphabet_size
-    
-    real(real64), intent(in) :: a
-    
-    log_fpa = log(alphabet_size * trigamma(alphabet_size*a + one) - trigamma(a + one))
-    
-  end function log_fpa
-
-  elemental real(real64) function log_fwa(alpha)
+  elemental real(real64) function log_weight(alpha)
     ! un-normalized weight for alpha in the integrals; prop. to p(alpha|x)
-    use dirichlet_mod, only: log_fpxa
+    use dirichlet_mod, only: log_pna, log_fpa
 
     real(real64), intent(in) :: alpha
+    
+    log_weight = log_fpa(alpha) + log_pna(alpha)
 
-    log_fwa = log_fpa(alpha) + log_fpxa(alpha)
-
-  end function log_fwa
+  end function log_weight
 
   subroutine compute_integration_range()
-    use dirichlet_mod, only: log_fpxa
+    use dirichlet_mod, only: log_pna
     
     integer(int32),parameter :: nx = 100
-    real(real64)             :: dx,largest,small_number
+    real(real64)             :: dx,largest
     real(real64)             :: xs(nx),fxs(nx)
     real(real64)             :: a1,a2,f,x
-    integer(int32)           :: i
-    
+    integer(int32)           :: i, counter, nbins
+
     largest = huge(dx)
-    small_number = 1.0e-30_real64
-    log_alpha1 = -20.0_real64
-    log_alpha2 = 10.0_real64
-
-    ! initialize amax
-    amax = 1.0_real64
-    log_fwa_amax = log_fwa(amax) ! log p(n|a_max)
-
-    ! set intervals equally spaced on log scale
-    dx = (log_alpha2 - log_alpha1) / (nx * 1.0_real64)
-    do i = 1,nx
-       xs(i) = log_alpha1 + (i - 0.5_real64) * dx
-    end do
-    xs = exp(xs)
     
-    fxs = log_fwa(xs)
-    ! find amax such that the weight in the integral - fwa(alpha) - is maximal
-    i = maxloc(fxs,1,fxs < largest)
-    amax = xs(i)
-    log_fwa_amax = log_fwa(amax)
+    ! initialize amax and integration range
+    log_alpha1 = log(1.e-8_real64)
+    log_alpha2 = log(1.e4_real64)
+    amax = 1.0_real64
+    lw_max = log_weight(amax)
+    
+    counter = 0
+    do
+       counter = counter + 1
+       
+       ! set intervals equally spaced on log scale
+       dx = (log_alpha2 - log_alpha1) / (nx * 1.0_real64)
+       do i = 1,nx
+          xs(i) = log_alpha1 + (i - 0.5_real64) * dx
+       end do
+       xs = exp(xs)
 
-    ! recompute fxs
-    fxs = exp(log_fwa(xs) - log_fwa_amax)
+       fxs = log_weight(xs)
+       ! find amax such that the alpha weight is maximal
+       i = maxloc(fxs, 1, fxs < largest)
+       amax = xs(i)
+       lw_max = log_weight(amax)
+
+       ! check the bins with weights > 0
+       fxs = exp(fxs - lw_max)
+       nbins = count(fxs > 0.0)
+       if (nbins > 1) exit
+       if (nbins == 1) then
+          log_alpha1 = log(amax) - dx
+          log_alpha2 = log(amax) + dx
+       end if
+    end do
 
     ! re-compute a reasonable integration range
-    log_alpha1 = log(minval(xs, fxs > small_number))
-    log_alpha2 = log(maxval(xs, fxs > small_number))
+    fxs = exp(log_weight(xs) - lw_max)
+    log_alpha1 = log(minval(xs, fxs > 0.0_real64)) - dx
+    log_alpha2 = log(maxval(xs, fxs > 0.0_real64)) + dx
+    
     dx = (log_alpha2 - log_alpha1) / (nx * 1.0_real64)
     do i = 1,nx
        xs(i) = log_alpha1 + (i - 0.5_real64) * dx
     end do
     xs = exp(xs)
     
-    fxs = log_fwa(xs)
-    amax = xs(maxloc(fxs,1,fxs < largest))
-    log_fwa_amax = log_fwa(amax)
-    
+    fxs = log_weight(xs)
+    ! find amax such that the alpha weight is maximal
+    i = maxloc(fxs, 1, fxs < largest)
+    amax = xs(i)
+    lw_max = log_weight(amax)
+
   end subroutine compute_integration_range
 
   real(real64) function m_func(x)
-    ! integrating after change of variable x = log(alpha) in the integral(s)
-    use dirichlet_mod, only: hdir
+    ! integrate over x = log(alpha)
+    use dirichlet_mod, only: h_bayes, integrand
 
     real(real64), intent(in) :: x
-    real(real64) :: a
+    real(real64) :: alpha
 
-    a = exp(x)
-    m_func = exp(log_fwa(a) - log_fwa_amax) * hdir(a) * a
+    alpha = exp(x)
+    m_func = integrand(alpha, lw_max, 1) * alpha
     
   end function m_func
 
   real(real64) function m2_func(x)
-    ! integrating after change of variable x = log(alpha) in the integral(s)
-    use dirichlet_mod, only: hdir
+    ! integrate over x = log(alpha)
+    use dirichlet_mod, only: h_bayes, integrand
 
     real(real64), intent(in) :: x
-    real(real64) :: a
+    real(real64) :: alpha
 
-    a = exp(x)
-    m2_func = exp(log_fwa(a) - log_fwa_amax) * hdir(a)**2 * a
+    alpha = exp(x)
+    m2_func = integrand(alpha, lw_max, 2) * alpha
     
   end function m2_func
 
   real(real64) function nrm_func(x)
-    ! integrating after change of variable x = log(alpha) in the integral(s)
-
+    ! integrate over x = log(alpha)
     real(real64), intent(in) :: x
-    real(real64) :: a
+    real(real64) :: alpha
 
-    a = exp(x)
-    nrm_func = exp(log_fwa(a) - log_fwa_amax)  * a
-    
+    alpha = exp(x)
+    nrm_func = exp(log_weight(alpha) - lw_max)  * alpha
+
   end function nrm_func
 
   subroutine hnsb(estimate,err_estimate) 
@@ -293,28 +342,43 @@ subroutine plugin(n,counts,estimate)
   real(real64)   :: ni,n_data
   integer(int32)              :: mi,nmax,err
   integer(int32), allocatable :: multi0(:)
+  logical :: multi = .false.
 
-  nbins = size(counts)
-  if (nbins == 1) then 
+  if (multi) then
+     ! using multiplicities 
+     nbins = size(counts)
+     if (nbins == 1) then 
+        estimate = 0.0_real64
+        return
+     end if
+     n_data = sum(counts)*1.0_real64
+     nmax = maxval(counts)
+     allocate(multi0(nmax),stat=err)
+     multi0 = 0
+     do i = 1,nbins
+        ni = counts(i)
+        if (ni == 0) cycle
+        multi0(ni) = multi0(ni) + 1
+     end do
      estimate = 0.0_real64
-     return
+     do i = 1,nmax
+        mi = multi0(i)
+        if (mi > 0) estimate = estimate - mi*i*log(i*1.0_real64)
+     end do
+     estimate = estimate / n_data + log(n_data)
+     deallocate(multi0)
+  else
+     ! standard implementation
+     nbins = size(counts)
+     if (nbins == 1) then 
+        estimate = 0.0_real64
+        return
+     end if
+     n_data = sum(counts)*1.0_real64
+     estimate = - sum(counts * log(counts*1.0_real64), counts>0)
+     estimate = estimate / n_data + log(n_data)
   end if
-  n_data = sum(counts)*1.0_real64
-  nmax = maxval(counts)
-  allocate(multi0(nmax),stat=err)
-  multi0 = 0
-  do i = 1,nbins
-     ni = counts(i)
-     if (ni == 0) cycle
-     multi0(ni) = multi0(ni) + 1
-  end do
-  estimate = 0.0_real64
-  do i = 1,nmax
-     mi = multi0(i)
-     if (mi > 0) estimate = estimate - mi*i*log(i*1.0_real64)
-  end do
-  estimate = estimate / n_data + log(n_data)
-  deallocate(multi0)
+
 
 end subroutine plugin
 
@@ -353,10 +417,10 @@ subroutine pseudo(n,counts,nc,alpha,estimate)
   end if
 
   nbins = size(counts)
-  if (nbins == 1) then 
-     estimate = 0.0_real64
-     return
-  end if
+!  if (nbins == 1) then 
+!     estimate = 0.0_real64
+!     return
+!  end if
   n_data = sum(counts)
   estimate = 0.0_real64
   do i = 1,nbins
@@ -377,7 +441,7 @@ subroutine dirichlet(n,counts,nc,alpha,estimate)
   ! posterior mean entropy (averaged over Dirichlet distribution) given alpha 
   use iso_fortran_env
   use dirichlet_mod, only: initialize_dirichlet, compute_multiplicities, dirichlet_finalize
-  use dirichlet_mod, only: hdir
+  use dirichlet_mod, only: h_bayes
   implicit none
 
   integer(int32), intent(in)  :: n
@@ -386,15 +450,15 @@ subroutine dirichlet(n,counts,nc,alpha,estimate)
   real(real64),   intent(in)  :: alpha
   real(real64),   intent(out) :: estimate
 
-  if (size(counts) == 1) then 
-     estimate = 0.0_real64
-     return
-  end if
+!  if (size(counts) == 1) then 
+!     estimate = 0.0_real64
+!     return
+!  end if
 
   call initialize_dirichlet(counts, nc)
   call compute_multiplicities(counts)
 
-  estimate = hdir(alpha)
+  estimate = h_bayes(alpha)
 
   call dirichlet_finalize()
 
@@ -413,13 +477,14 @@ subroutine nsb(n,counts,nc,estimate,err_estimate)
   real(real64),   intent(out) :: estimate
   real(real64),   intent(out) :: err_estimate
 
-  if (size(counts) == 1) then 
-     estimate = 0.0_real64
-     err_estimate = 0.0_real64
-     return
-  end if
+!  if (size(counts) == 1) then 
+!     estimate = 0.0_real64
+!     err_estimate = 0.0_real64
+!     return
+!  end if
 
   call initialize_dirichlet(counts, nc)
+
   call compute_multiplicities(counts)
 
   call compute_integration_range()